+++ docs/conf.py
@@ -59,7 +59,9 @@ if "dev" in release:
 exclude_patterns = ["_build"]
 
 nitpick_ignore = [
+    ("py:class", "AbstractAsyncContextManager"),
     ("py:class", "aiohttp.web_request.Request"),
+    ("py:class", "FastAPI"),
 ]
 
 # If true, '()' will be appended to :func: etc. cross-reference text.